+++ progs/lbfgsb_fortran.c
@@ -32,13 +32,13 @@
 #include <math.h>
 #include <string.h>
 
-typedef long int integer;
-typedef unsigned long int uinteger;
+typedef int integer;
+typedef unsigned int uinteger;
 typedef char *address;
 typedef short int shortint;
 typedef float real;
 typedef double doublereal;
-typedef long int logical;
+typedef int logical;
 typedef short int shortlogical;
 typedef char logical1;
 typedef char integer1;
